+++ media/utils/addJiraTicketsToDailySummary.js
@@ -1,11 +1,14 @@
 const addJiraTicketsToDailySummary = (jiraTickets) => {
   $("#jiraHolder").append(`<h3>Assigned Jira Tickets</h3>`);
   jiraTickets.forEach((ticket) => {
+    $("#jiraHolder").append(`
     <div class="Box-header">
-      <h5 class="Box-title Truncate">${ticket?.key}</h5>
-      <span style="color: red">${ticket?.fields?.status?.name}</span>
-      <p>${ticket?.fields?.summary}</p>
-    </div>;
+        <h5 class="Box-title Truncate">${ticket?.key}</h5>
+        <span style="color: red">${ticket?.fields?.status?.name}</span>
+        <p>${ticket?.fields?.summary}</p>
+    </div>
+    <br/>
+    `);
   });
 };