+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [1.6.1]
 - Get code context button now retrieves the most relevant Jira ticket, if the user is authenticated with Jira
 - Watermelon is now an authentication provider, to allow us to connect to all your services
+- Allows the user to deactivate telemetry through settings
 
 ## [1.6.0]
 - Watermelon is now an authentication provider, to allow us to connect to all your services